refactor stars composable
