+++ app/src/main/java/uba/fi/goodreads/presentation/bookInfo/composables/ratingStars.kt
@@ -1,10 +1,9 @@
-package uba.fi.goodreads.ui.components.ratingStars
+package uba.fi.goodreads.presentation.bookInfo.composables
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.material.icons.Icons
@@ -13,35 +12,27 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
-import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
+import uba.fi.goodreads.presentation.bookInfo.BookInfoScreenPreviewParameterProvider
+import uba.fi.goodreads.presentation.bookInfo.BookInfoUIState
 
 @Composable
 fun RatingScreen(
-    ratingVewModel: RatingViewModel = viewModel()
+    screenState: BookInfoUIState
 ) {
-    val rating by ratingVewModel.rating.collectAsState()
-
     Column(
         horizontalAlignment = Alignment.CenterHorizontally,
         verticalArrangement = Arrangement.Center,
     ) {
         Text(text = "My rate", style = MaterialTheme.typography.labelMedium)
-
         RatingStars(
-            rating = rating,
-            onRatingChange = { newRating -> ratingVewModel.updateRating(newRating) }
+            rating = screenState.rating,
+            onRatingChange = {newRating -> screenState.rating = newRating}
         )
     }
 }
@@ -51,7 +42,7 @@ fun RatingStars(
     rating: Int,
     onRatingChange: (Int) -> Unit
 ) {
-    val validRating = rating.coerceIn(0, 5)
+    var validRating = rating.coerceIn(0, 5)
     Row {
         (1..5).forEach { index ->
             Icon(
@@ -67,17 +58,13 @@ fun RatingStars(
     }
 }
 
-class RatingStarsPreviewParameterProvider : PreviewParameterProvider<Int> {
-    override val values: Sequence<Int> = sequenceOf(0, 1, 2, 3, 4, 5)
-}
+
 
 @Preview(showBackground = true)
 @Composable
-fun RatingStarsPreview(
-    @PreviewParameter(RatingStarsPreviewParameterProvider::class) rating: Int
-) {
+fun RatingStarsPreview() {
     RatingStars(
-        rating = rating,
+        rating = 3,
         onRatingChange = {}
     )
 }
@@ -85,8 +72,7 @@ fun RatingStarsPreview(
 @Preview(showBackground = true)
 @Composable
 fun RatingScreenPreview(
-    @PreviewParameter(RatingStarsPreviewParameterProvider::class) rating: Int
+    @PreviewParameter(BookInfoScreenPreviewParameterProvider::class) state: BookInfoUIState
 ) {
-    val previewViewModel = RatingViewModel().apply { updateRating(rating) }
-    RatingScreen(previewViewModel)
+    RatingScreen(screenState = state)
 }
